refactor: Consolidate single instance services into PosixSingleInstanceService
Remove three duplicate implementations (Avalonia, macOS, Linux) and
replace with a single PosixSingleInstanceService in TerminalHost.Posix.
Avalonia app now uses it directly via ISingleInstanceService.
